requirements.txt
# AidSight Python 依赖

# 核心依赖
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=5.4.0

# MindSpore (需要根据实际硬件安装对应版本)
# mindspore>=2.2.0  # CPU 版本
# 昇腾版本需要从官方渠道安装

# MindYOLO (模型训练)
# mindyolo>=0.1.0  # 从 GitHub 安装

# 推理引擎 (根据实际部署环境选择)
# onnxruntime>=1.12.0        # ONNX 推理
# onnxruntime-gpu>=1.12.0    # ONNX GPU 推理

# 音频处理 (可选)
# pygame>=2.0.0              # 音频播放
# pyttsx3>=2.90              # TTS 文本转语音
# gTTS>=2.2.0                # Google TTS

# 蓝牙支持 (可选，Linux)
# pybluez>=0.23              # 蓝牙库

# 开发工具
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# 实用工具
tqdm>=4.62.0
colorama>=0.4.4
pillow>=9.0.0
